Rename repo to java-api-extensions, artifact id to opentracing-api-extensions and package name to io.opentracing.contrib.api
