+++ README.md
@@ -1,13 +1,18 @@
 [![Build Status][ci-img]][ci] [![Released Version][maven-img]][maven]
 
-# OpenTracing Observer
+# OpenTracing API Extensions
 
-This project provides an observer API for the OpenTracing Tracer.
+This project provides API extensions to the core OpenTracing APIs.
+
+## Observer
+
+TODO: Provide some code snippets for use of Observer API.
+
 
 ## Release
 Follow instructions in [RELEASE](RELEASE.md)
 
-   [ci-img]: https://travis-ci.org/opentracing-contrib/java-observer.svg?branch=master
-   [ci]: https://travis-ci.org/opentracing-contrib/java-observer
-   [maven-img]: https://img.shields.io/maven-central/v/io.opentracing.contrib/opentracing-observer.svg?maxAge=2592000
-   [maven]: http://search.maven.org/#search%7Cga%7C1%7Copentracing-observer
+   [ci-img]: https://travis-ci.org/opentracing-contrib/java-api-extensions.svg?branch=master
+   [ci]: https://travis-ci.org/opentracing-contrib/java-api-extensions
+   [maven-img]: https://img.shields.io/maven-central/v/io.opentracing.contrib/opentracing-api-extensions.svg?maxAge=2592000
+   [maven]: http://search.maven.org/#search%7Cga%7C1%7Copentracing-api-extensions
