+++ opentracing-api-extensions/pom.xml
@@ -17,12 +17,12 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <artifactId>opentracing-observer-parent</artifactId>
+    <artifactId>opentracing-api-extensions-parent</artifactId>
     <groupId>io.opentracing.contrib</groupId>
     <version>0.0.1-SNAPSHOT</version>
   </parent>
 
-  <artifactId>opentracing-observer</artifactId>
+  <artifactId>opentracing-api-extensions</artifactId>
 
   <dependencies>
     <dependency>
