+++ opentracing-api-extensions/src/main/java/io/opentracing/contrib/api/SpanObserver.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing.contrib.observer;
+package io.opentracing.contrib.api;
 
 import java.util.Map;
